unbreak 'required input' validation check in survey form
Before, we listen for the 'click' event of the submit button associated with the form.
Sadly, this had the side effect of disabling the 'required input' checking specified in our html:

<form id="surveyForm">
  <legend><h3><strong>${json.title}</strong></legend>
  <input name="name" type="text" placeholder="${json.namePlaceholder}" required></input>
                                                                       --------
  ..
  <hr>
  ..
</form>

Now we listen for the 'submit' event of the form itself an regain the browser-provided validation checking.
To summarize:

:-(  delegate(document, "click", "#submit-survey", postSurveyForm);
:-)  delegate(document, "submit", "#surveyForm",   postSurveyForm);