fix issue with unresponsive click handler on iOS safari
A friend of mine tested this application on iOS and
the app selection 'onchange' event did not trigger
the bring-up of the expected demo application.

It just highlighted the selection and sat there on the
Similarity Engine landing page.

Same deal with the results dismissal 'X'.  It would change
color to 'red' but not actually trigger the code that should
hide the results div.

I'm hoping the stackoverflow fu here will do the trick:

https://stackoverflow.com/questions/24077725/mobile-safari-sometimes-does-not-trigger-the-click-event

I believe this is an edgecase side-effect of me changing out the contents of the <body>
dynamically.